Ignore an OpenAPI JSON update if there's no type changes
This works around the OpenAPI JSON now always being "updated" with the
new version, as per https://github.com/OpusDNS/opusdns-api/pull/1070

So, instead, it's not committing anything unless the types have changed.

That might not _always_ be correct (if the OpenAPI metadata changed?),
but it should be correct in the vast majority of cases so let's start
with it to avoid noisy commits every time someone makes a change to
backend.