Use "unboxed" version of TypeID
https://github.com/jetify-com/opensource/tree/main/typeid/typeid-js/src/unboxed

We've been using TypeID wrong the whole time. The default one is
class-based. Which is usually what you want. *Except* in
serializing/deserializing context, like we have here. Because we're
dealing with strings, we can say the type of `user_id` is
`TypeID<'user'>`, but... that does not mean anything. Underneath,
`typeof user_id` will be `string`, not `TypeID<'user'>`.

To fix this, we need to switch to "unboxed" version of TypeID, which
_extends_ the `string` type. It gives us the same type security (and
uses the boxed version of TypeID underneath), while playing very nicely
with strings from JSON, etc.