Prevent concurrent runs of the job to avoid failures
The job is triggered by new merges on the main branch in a different repository - and that sends an event to this repository to start a new job run. But if there's two merges one right after the other, one job could start before the other finishes and that prevents the second one from finishing, because it is prevented from commiting the changes due to the repository having changed in the meantime.