Switch to using the scripts in @opusdns/api-spec-ts-generator
https://github.com/OpusDNS/api-spec-ts-generator

This allows this repository to be simply about the generated artifacts,
related to OpusDNS. And if someone wants to use the scripts used to
generate these types (which should not be OpusDNS-specific), then those
are now available in a separate repository.
It will also make it easier to make updates to the scripts, maintain
them, etc.