Improve change detection to TS files and API schema
The previous approach was not working well - it was ignoring the YAML
file, because the `version` would always change between runs. So it'd
end up with false positives. That's why it was only looking if the
generated TypeScript types changed. But that was not always enough.
Sometimes, there were changes in the YAML file that were only about
documentation, that was not directly affecting the TS types.

Now, this improved workflow should use the "old" (current) version of
the schema file, so that git should not show any changes for that field.
And now we can simply generate everything, see if anything changed. If
it did, we then use the unmodified API YAML file to generate once again
everything, this time with the correct/new version. Not the most
efficient thing, but it takes literally a second and does not run often.