Use sed for setting the temporary version
This is because apparently the generated OpenAPI YAML file is being
reformatted by yq. Specifically, the indentation for lists?
This avoids yq introducing those changes and just uses `sed` for simple
search and replace.
Ideally, we'd fix it upstream, but it seems this is how the upstream's
upstream OpenAPI generator works, so that might not be possible
short-term.